# Enable column projection for `DataFrame.index` and further optimize `DataFrame.__len__`