@rjzamora 2023-03-15
**Note**: It seems that the `__len__` optimization introduced here requires all [`DataFrameIOFunction`](https://github.com/dask/dask/blob/8adf6ba81fb93e152243f202045c3ca052c331de/dask/dataframe/io/utils.py#L224) to produce a `pd/cudf.DataFrame` object with a consistent length, regardless of the column selection. In `main`, it is **not** always the case that specifying `columns=[]` will produce a `DataFrame` of non-zero length.

This PR includes a few "bug-fixes" related to this empty-columns problem. However, it is not clear that it is technically a "bug" to return an empty `DataFrame` when there are no columns, and so it is probably a bit too fragile to make an empty column selection inside `DataFrame.index`.